fix apim using public

+++ iac/helm-chart/chart/deploy.ps1
@@ -1,6 +1,6 @@
 param(
     [string]$ChartName = "logcorner.edusync.speech",
-    [string]$IMAGE_TAG = "1118",
+    [string]$IMAGE_TAG = "1121",
     [string]$NAMESPACE = "ingress-nginx",
     [string]$RESOURCE_GROUP_NAME = "rg-edusync-dev",
     [string]$CLUSTER_NAME = "aks-edusync-dev",
@@ -18,7 +18,7 @@ kubectl get deployments --all-namespaces=true
 az network vnet check-ip-address --name $VNET_NAME -g $RESOURCE_GROUP_NAME --ip-address $PRIVATE_IP
 
 # Add the ingress-nginx Helm repository if not already added
-Write-Output "Adding the ingress-nginx repository..."
+Write-Host "Adding the ingress-nginx repository..." -ForegroundColor Green
  # Add the ingress-nginx repository
 helm repo add ingress-nginx https://kubernetes.github.io/ingress-nginx
 
@@ -26,7 +26,7 @@ helm repo add ingress-nginx https://kubernetes.github.io/ingress-nginx
 helm repo update
 
 # Deploy the NGINX ingress controller with an internal load balancer
-Write-Output "Deploying the NGINX ingress controller..."
+Write-Host "Deploying the NGINX ingress controller..." -ForegroundColor Green
 # # Use Helm to deploy an NGINX ingress controllerusing static private IP address
 #  helm upgrade --install ingress-nginx ingress-nginx/ingress-nginx  `
 #   --namespace $NAMESPACE  `
@@ -44,31 +44,43 @@ helm upgrade --install ingress-nginx ingress-nginx/ingress-nginx  `
     
 # Test an internal IP address
 # Monitor the ingress service
-Write-Output "Monitoring ingress-nginx-controller service..."
+Write-Host "Monitoring ingress-nginx-controller service..." -ForegroundColor Green
 #kubectl  get services -o wide -w ingress-nginx-controller -n $NAMESPACE
 kubectl get service ingress-nginx-controller --namespace  $NAMESPACE
 
 # Deploy an additional Helm chart (logcorner-command)
-Write-Output "Deploying logcorner-command chart..."
+Write-Host "Deploying logcorner-command chart..." -ForegroundColor Green
 # Change to the correct directory (up one level)
 # Define the full path to the Helm chart directory
 #$ChartName = "logcorner.edusync.speech"
-kubectl delete pod curl-test --namespace  helm
+
+
+# kubectl delete pod curl-test --namespace  helm
+
+$status = kubectl get pod curl-test --namespace helm -o jsonpath='{.status.phase}'
+if ($status -ne "Running") {
+    kubectl delete pod curl-test --namespace helm
+}
+
 
 # helm upgrade --install logcorner-command  $ChartName
 helm upgrade --install logcorner-command $ChartName `
     --set global.tag=$IMAGE_TAG
 
 $NAMESPACE="default"
-write-host "Waiting for the logcorner-command pod to be ready..."
+write-host "Waiting for the logcorner-command pod to be ready... " -ForegroundColor Green
+kubectl wait --for=condition=ready pod -l app=logcorner-command-http-api --timeout=300s
+
 kubectl get pods --namespace  $NAMESPACE
 
 $NAMESPACE="ingress-nginx"
 kubectl get service ingress-nginx-controller --namespace  $NAMESPACE
 
+Write-Host "Getting private ip of ingress controller..." -ForegroundColor Green
 
 $PRIVATE_IP=kubectl get svc ingress-nginx-controller -n ingress-nginx -o jsonpath='{.status.loadBalancer.ingress[0].ip}'
 
+Write-Host "Adding dns record to private dns zone..." -ForegroundColor Green
 
 # az network private-dns record-set a add-record  `
 #   -g $RESOURCE_GROUP_NAME  `
@@ -87,21 +99,29 @@ $existingRecord = az network private-dns record-set a show `
   --output tsv
 
 if (-not $existingRecord) {
-    Write-Host "Record does not exist. Adding A record..."
+    Write-Host "Record does not exist. Adding A record..." -ForegroundColor Green
     az network private-dns record-set a add-record `
       -g $RESOURCE_GROUP_NAME `
       -z $PrivateDnsZoneName  `
       -n $RecordName `
       -a $PRIVATE_IP
-    Write-Host "A record added: $RecordName -> $PRIVATE_IP"
+    Write-Host "A record added: $RecordName -> $PRIVATE_IP" -ForegroundColor Green
 } else {
-    Write-Host "Record already exists: $RecordName -> $PRIVATE_IP"
+    Write-Host "Record already exists: $RecordName -> $PRIVATE_IP" -ForegroundColor Yellow
 }
 
+Write-Host "Running dns resolution between $PrivateDnsZoneName and $PRIVATE_IP  ..." -ForegroundColor Green
 kubectl exec -it curl-test -n helm -- nslookup ingress.cloud-devops-craft.com
+
+Write-Host "Running test ..." -ForegroundColor Green
+
+Write-Host "Getting all WeatherForecast ..." -ForegroundColor Green
 
 kubectl exec -it curl-test -n helm -- curl http://ingress.cloud-devops-craft.com/aks-command-api/WeatherForecast
 
+Write-Host "Getting  WeatherForecast by id ..." -ForegroundColor Green
+
+kubectl exec -it curl-test -n helm -- curl http://ingress.cloud-devops-craft.com/aks-command-api/WeatherForecast/1
 
 # kubectl exec -it curl-test -n helm -- curl http://$PRIVATE_IP/aks-command-api/WeatherForecast
